+++ Assets/Resource/Shader/MiniMap/MiniMapC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c2cfc8cca64f654fa9b2fa8cc525c2e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
